Refactor transcription tests and remove deprecated test file
This commit refactors the transcription tests by consolidating them into a new file, `test_transcription.py`, while removing the old `test_transcription.py` file. The new test suite includes comprehensive tests for the `TranscriptionConfig`, `TranscriptionResult`, and transcriber classes, ensuring robust coverage of functionality and edge cases. Additionally, minor adjustments were made to improve test organization and clarity.
